Refactor `Contact` section
